Implemented Conversation Mode.
Skills can now access the ThreadManager from the overseer.
They can create Threads and store data into the hub, which will call the skill handler for this interaction when it receives an answer (see quizz skill example).
Threads are answered at `/converse` endpoint (so adapters must make the difference between a conversation and a direct command/message). Messages returnes by skills contains the thread ID and a `interactive: true` variable, so it's easy to know when a conversation is currently active.