Added hook management.
Skills may now register for a hook with a connector. If the connector finalizes the hook, the skill will be able to use this pipe to display messages to the hooked connector (the connector may then restrict to a unique room or do whatever it wants to).
See `alarm` skill for an example of a hook with a simple alarm skill.